refactor: move workspace-content identity spec into the source-snapshot contract
Lab materialization verification hinges on the controller and runner computing
an identical workspace content hash, which requires an identical algorithm
identity -- the marker string naming the hashing rules (traversal version +
permission policy). That spec (the permission-policy names, the default policy,
and the policy -> algorithm-marker mapping) lived in homeboy-lab-runner's
filesystem-walking module, buried next to the traversal behavior. The
controller-declare path (lab/offload/metadata.rs) and the runner-verify path
(workspace/provenance.rs) both derive the marker from it, so it is a shared
contract, not runner behavior.

Move the pure identity spec -- WORKSPACE_CONTENT_PERMISSION_* constants,
WORKSPACE_CONTENT_DEFAULT_PERMISSION_POLICY, workspace_content_hash_algorithm(),
and the WorkspaceContentManifest[Entry] types -- DOWN into the leaf
homeboy-source-snapshot-contract crate as workspace_content_identity, with the
v1/v2/v3 algorithm lineage documented in one place. lab-runner re-exports it so
existing paths resolve unchanged; the filesystem traversal that *applies* the
spec (mode reading, byte hashing, exclude matching) stays in the runner.

This makes 'what a content identity means' a single documented spec in a leaf
crate that any side can depend on without pulling in the runner -- part of
stabilizing the recurring Lab snapshot-identity churn by giving identity one
authoritative definition.

Verified: contract 3 spec tests pass, lab-runner --lib clean, all content-hash/
policy/algorithm snapshot tests pass. (git_backed_snapshot_uses_runner_exact_sha
fails identically on origin/main -- pre-existing git/SSH env test.)
